Enum hydroflow_plus::bincode::ErrorKind

pub enum ErrorKind {
    Io(Error),
    InvalidUtf8Encoding(Utf8Error),
    InvalidBoolEncoding(u8),
    InvalidCharEncoding,
    InvalidTagEncoding(usize),
    DeserializeAnyNotSupported,
    SizeLimit,
    SequenceMustHaveLength,
    Custom(String),
}

The kind of error that can be produced during a serialization or deserialization.

Variants

Io(Error)

If the error stems from the reader/writer that is being used during (de)serialization, that error will be stored and returned here.

InvalidUtf8Encoding(Utf8Error)

Returned if the deserializer attempts to deserialize a string that is not valid utf8

InvalidBoolEncoding(u8)

Returned if the deserializer attempts to deserialize a bool that was not encoded as either a 1 or a 0

InvalidCharEncoding

Returned if the deserializer attempts to deserialize a char that is not in the correct format.

InvalidTagEncoding(usize)

Returned if the deserializer attempts to deserialize the tag of an enum that is not in the expected ranges

DeserializeAnyNotSupported

Serde has a deserialize_any method that lets the format hint to the object which route to take in deserializing.

SizeLimit

If (de)serializing a message takes more than the provided size limit, this error is returned.

SequenceMustHaveLength

Bincode can not encode sequences of unknown length (like iterators).

Custom(String)

A custom error message from Serde.