Function hydroflow_plus::bincode::deserialize_from

pub fn deserialize_from<R, T>(reader: R) -> Result<T, Box<ErrorKind>>
where
    R: Read,
    T: DeserializeOwned,

Deserializes an object directly from a Reader using the default configuration.

If this returns an Error, reader may be in an invalid state.

Warning: the default configuration used by this function is not the same as that used by the DefaultOptions struct. See the config module for more details