Struct hydroflow_plus::bincode::Serializer

pub struct Serializer<W, O>
where
    O: Options,
{ /* private fields */ }

An Serializer that encodes values directly into a Writer.

The specified byte-order will impact the endianness that is used during the encoding.

This struct should not be used often. For most cases, prefer the encode_into function.

Implementations

impl<W, O> Serializer<W, O>
where W: Write, O: Options,

pub fn new(w: W, options: O) -> Serializer<W, O>

Creates a new Serializer with the given Writer.