Trait hydroflow_plus::bincode::Options

pub trait Options: InternalOptions + Sized {

    // Provided methods
    fn with_no_limit(self) -> WithOtherLimit<Self, Infinite> { ... }
    fn with_limit(self, limit: u64) -> WithOtherLimit<Self, Bounded> { ... }
    fn with_little_endian(self) -> WithOtherEndian<Self, LittleEndian> { ... }
    fn with_big_endian(self) -> WithOtherEndian<Self, BigEndian> { ... }
    fn with_native_endian(self) -> WithOtherEndian<Self, NativeEndian> { ... }
    fn with_varint_encoding(self) -> WithOtherIntEncoding<Self, VarintEncoding> { ... }
    fn with_fixint_encoding(self) -> WithOtherIntEncoding<Self, FixintEncoding> { ... }
    fn reject_trailing_bytes(self) -> WithOtherTrailing<Self, RejectTrailing> { ... }
    fn allow_trailing_bytes(self) -> WithOtherTrailing<Self, AllowTrailing> { ... }
    fn serialize<S>(self, t: &S) -> Result<Vec<u8>, Box<ErrorKind>>
       where S: Serialize + ?Sized { ... }
    fn serialized_size<T>(self, t: &T) -> Result<u64, Box<ErrorKind>>
       where T: Serialize + ?Sized { ... }
    fn serialize_into<W, T>(self, w: W, t: &T) -> Result<(), Box<ErrorKind>>
       where W: Write,
             T: Serialize + ?Sized { ... }
    fn deserialize<'a, T>(self, bytes: &'a [u8]) -> Result<T, Box<ErrorKind>>
       where T: Deserialize<'a> { ... }
    fn deserialize_seed<'a, T>(
        self,
        seed: T,
        bytes: &'a [u8]
    ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
       where T: DeserializeSeed<'a> { ... }
    fn deserialize_from<R, T>(self, reader: R) -> Result<T, Box<ErrorKind>>
       where R: Read,
             T: DeserializeOwned { ... }
    fn deserialize_from_seed<'a, R, T>(
        self,
        seed: T,
        reader: R
    ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
       where R: Read,
             T: DeserializeSeed<'a> { ... }
    fn deserialize_from_custom<'a, R, T>(
        self,
        reader: R
    ) -> Result<T, Box<ErrorKind>>
       where R: BincodeRead<'a>,
             T: DeserializeOwned { ... }
    fn deserialize_from_custom_seed<'a, R, T>(
        self,
        seed: T,
        reader: R
    ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
       where R: BincodeRead<'a>,
             T: DeserializeSeed<'a> { ... }

}

A configuration builder trait whose options Bincode will use while serializing and deserializing.

Options

Endianness: The endianness with which multi-byte integers will be read/written. default: little endian

Limit: The maximum number of bytes that will be read/written in a bincode serialize/deserialize. default: unlimited

Int Encoding: The encoding used for numbers, enum discriminants, and lengths. default: varint

Trailing Behavior: The behavior when there are trailing bytes left over in a slice after deserialization. default: reject

Byte Limit Details

The purpose of byte-limiting is to prevent Denial-Of-Service attacks whereby malicious attackers get bincode deserialization to crash your process by allocating too much memory or keeping a connection open for too long.

When a byte limit is set, bincode will return Err on any deserialization that goes over the limit, or any serialization that goes over the limit.

Provided Methods

fn with_no_limit(self) -> WithOtherLimit<Self, Infinite>

Sets the byte limit to be unlimited. This is the default.

fn with_limit(self, limit: u64) -> WithOtherLimit<Self, Bounded>

Sets the byte limit to limit.

fn with_little_endian(self) -> WithOtherEndian<Self, LittleEndian>

Sets the endianness to little-endian This is the default.

fn with_big_endian(self) -> WithOtherEndian<Self, BigEndian>

Sets the endianness to big-endian

fn with_native_endian(self) -> WithOtherEndian<Self, NativeEndian>

Sets the endianness to the the machine-native endianness

fn with_varint_encoding(self) -> WithOtherIntEncoding<Self, VarintEncoding>

Sets the length encoding to varint

fn with_fixint_encoding(self) -> WithOtherIntEncoding<Self, FixintEncoding>

Sets the length encoding to be fixed

fn reject_trailing_bytes(self) -> WithOtherTrailing<Self, RejectTrailing>

Sets the deserializer to reject trailing bytes

fn allow_trailing_bytes(self) -> WithOtherTrailing<Self, AllowTrailing>

Sets the deserializer to allow trailing bytes

fn serialize<S>(self, t: &S) -> Result<Vec<u8>, Box<ErrorKind>>
where S: Serialize + ?Sized,

Serializes a serializable object into a Vec of bytes using this configuration

fn serialized_size<T>(self, t: &T) -> Result<u64, Box<ErrorKind>>
where T: Serialize + ?Sized,

Returns the size that an object would be if serialized using Bincode with this configuration

fn serialize_into<W, T>(self, w: W, t: &T) -> Result<(), Box<ErrorKind>>
where W: Write, T: Serialize + ?Sized,

Serializes an object directly into a Writer using this configuration

If the serialization would take more bytes than allowed by the size limit, an error is returned and no bytes will be written into the Writer

fn deserialize<'a, T>(self, bytes: &'a [u8]) -> Result<T, Box<ErrorKind>>
where T: Deserialize<'a>,

Deserializes a slice of bytes into an instance of T using this configuration

fn deserialize_seed<'a, T>( self, seed: T, bytes: &'a [u8] ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
where T: DeserializeSeed<'a>,

Deserializes a slice of bytes with state seed using this configuration.

fn deserialize_from<R, T>(self, reader: R) -> Result<T, Box<ErrorKind>>
where R: Read, T: DeserializeOwned,

Deserializes an object directly from a Reader using this configuration

If this returns an Error, reader may be in an invalid state.

fn deserialize_from_seed<'a, R, T>( self, seed: T, reader: R ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
where R: Read, T: DeserializeSeed<'a>,

Deserializes an object directly from a Reader with state seed using this configuration

If this returns an Error, reader may be in an invalid state.

fn deserialize_from_custom<'a, R, T>( self, reader: R ) -> Result<T, Box<ErrorKind>>
where R: BincodeRead<'a>, T: DeserializeOwned,

Deserializes an object from a custom BincodeReader using the default configuration. It is highly recommended to use deserialize_from unless you need to implement BincodeRead for performance reasons.

If this returns an Error, reader may be in an invalid state.

fn deserialize_from_custom_seed<'a, R, T>( self, seed: T, reader: R ) -> Result<<T as DeserializeSeed<'a>>::Value, Box<ErrorKind>>
where R: BincodeRead<'a>, T: DeserializeSeed<'a>,

Deserializes an object from a custom BincodeReader with state seed using the default configuration. It is highly recommended to use deserialize_from unless you need to implement BincodeRead for performance reasons.

If this returns an Error, reader may be in an invalid state.

Object Safety

This trait is not object safe.