Struct hydroflow_plus::bytes::buf::Writer

pub struct Writer<B> { /* private fields */ }

A BufMut adapter which implements io::Write for the inner value.

This struct is generally created by calling writer() on BufMut. See documentation of writer() for more details.

Implementations

impl<B> Writer<B>where B: BufMut,

pub fn get_ref(&self) -> &B

Gets a reference to the underlying BufMut.

It is inadvisable to directly write to the underlying BufMut.

Examples

use bytes::BufMut;

let buf = Vec::with_capacity(1024).writer();

assert_eq!(1024, buf.get_ref().capacity());

pub fn get_mut(&mut self) -> &mut B

Gets a mutable reference to the underlying BufMut.

It is inadvisable to directly write to the underlying BufMut.

Examples

use bytes::BufMut;

let mut buf = vec![].writer();

buf.get_mut().reserve(1024);

assert_eq!(1024, buf.get_ref().capacity());

pub fn into_inner(self) -> B

Consumes this Writer, returning the underlying value.

Examples

use bytes::BufMut;
use std::io;

let mut buf = vec![].writer();
let mut src = &b"hello world"[..];

io::copy(&mut src, &mut buf).unwrap();

let buf = buf.into_inner();
assert_eq!(*buf, b"hello world"[..]);