Trait hydroflow_plus::internal::syn::parse::Parse

pub trait Parse: Sized {
    // Required method
    fn parse(input: &ParseBuffer<'_>) -> Result<Self, Error>;
}

Parsing interface implemented by all types that can be parsed in a default way from a token stream.

Refer to the module documentation for details about implementing and using the Parse trait.

Required Methods

fn parse(input: &ParseBuffer<'_>) -> Result<Self, Error>

Object Safety

This trait is not object safe.