Enum hydroflow_plus::ir::HfPlusNode

pub enum HfPlusNode {

    Placeholder,
    Persist(Box<HfPlusNode>),
    Delta(Box<HfPlusNode>),
    Source {
        source: HfPlusSource,
        produces_delta: bool,
        location_id: usize,
    },
    CycleSource {
        ident: Ident,
        location_id: usize,
    },
    Tee {
        inner: Rc<RefCell<HfPlusNode>>,
    },
    Union(Box<HfPlusNode>, Box<HfPlusNode>),
    CrossProduct(Box<HfPlusNode>, Box<HfPlusNode>),
    Join(Box<HfPlusNode>, Box<HfPlusNode>),
    Difference(Box<HfPlusNode>, Box<HfPlusNode>),
    AntiJoin(Box<HfPlusNode>, Box<HfPlusNode>),
    Map {
        f: DebugExpr,
        input: Box<HfPlusNode>,
    },
    PipelineOp {
        kind: &'static str,
        gen_pipeline: DebugPipelineFn,
        input: Box<HfPlusNode>,
    },
    Network {
        to_location: usize,
        serialize_pipeline: Option<Pipeline>,
        sink_expr: DebugExpr,
        source_expr: DebugExpr,
        deserialize_pipeline: Option<Pipeline>,
        input: Box<HfPlusNode>,
    },

}

An intermediate node in a Hydroflow+ graph, which consumes data from upstream nodes and emits data to downstream nodes.

Variants

Placeholder

Persist(Box<HfPlusNode>)

Delta(Box<HfPlusNode>)

Source

Fields

source: HfPlusSource

produces_delta: bool

location_id: usize

CycleSource

Fields

ident: Ident

location_id: usize

Tee

Fields

inner: Rc<RefCell<HfPlusNode>>

Union(Box<HfPlusNode>, Box<HfPlusNode>)

CrossProduct(Box<HfPlusNode>, Box<HfPlusNode>)

Join(Box<HfPlusNode>, Box<HfPlusNode>)

Difference(Box<HfPlusNode>, Box<HfPlusNode>)

AntiJoin(Box<HfPlusNode>, Box<HfPlusNode>)

Map

Fields

f: DebugExpr

input: Box<HfPlusNode>

PipelineOp

Fields

kind: &'static str

gen_pipeline: DebugPipelineFn

input: Box<HfPlusNode>

A temporary IR node for a custom pipeline, which eventually will be replaced by a proper node.

Network

Fields

to_location: usize

serialize_pipeline: Option<Pipeline>

sink_expr: DebugExpr

source_expr: DebugExpr

deserialize_pipeline: Option<Pipeline>

input: Box<HfPlusNode>

Implementations

impl HfPlusNode

pub fn emit( self, graph_builders: &mut BTreeMap<usize, FlatGraphBuilder>, built_tees: &mut HashMap<*const RefCell<HfPlusNode>, (Ident, usize, bool)>, next_stmt_id: &mut usize ) -> (Ident, usize, bool)