Struct hydroflow_plus::lang::diagnostic::Diagnostic

pub struct Diagnostic<S = Span> {
    pub span: S,
    pub level: Level,
    pub message: String,
}

Diagnostic. A warning or error (or lower Level) with a message and span. Shown by IDEs usually as a squiggly red or yellow underline.

Must call Diagnostic::emit or manually emit the output of Diagnostic::to_tokens for the diagnostic to show up.

Fields
span: S

Span (source code location).

level: Level

Severity level.

message: String

Human-readable message.

Implementations

impl<S> Diagnostic<S>

pub fn is_error(&self) -> bool

If this diagnostic’s level is Level::Error.

impl Diagnostic

pub fn spanned( span: Span, level: Level, message: impl Into<String> ) -> Diagnostic

Create a new diagnostic from the given span, level, and message.

pub fn emit(&self)

Emit the diagnostic. Only works from the proc_macro context. Does not work outside of that e.g. in normal runtime execution or in tests.

pub fn to_tokens(&self) -> TokenStream

Used to emulate Diagnostic::emit by turning this diagnostic into a properly spanned TokenStream that emits an error with this diagnostic’s message.

pub fn to_serde(&self) -> Diagnostic<SerdeSpan>

Converts this into a serializable and deserializable Diagnostic. Span information is converted into SerdeSpan which keeps the span info but cannot be plugged into or emitted through the Rust compiler’s diagnostic system.