Module hydroflow_plus::lang::graph::ops

Hydroflow’s operators

Structs

FlowPropArgs

Arguments provided to OperatorConstraints::flow_prop_fn.

OperatorConstraints

An instance of this struct represents a single hydroflow operator.

OperatorWriteOutput

The code generated and returned by a OperatorConstraints::write_fn.

WriteContextArgs

Context arguments provided to OperatorConstraints::write_fn.

Enums

DelayType

The delay (soft barrier) type, for each input to an operator if needed.

OperatorCategory

Operator categories, for docs.

Persistence

Persistence lifetimes: 'tick, 'static, or 'mutable.

PortListSpec

Specification of the named (or unnamed) ports for an operator’s inputs or outputs.

Constants

IDENTITY_WRITE_FN

OperatorConstraints::write_fn for unary identity operators.

JOIN_CROSS_JOIN_FLOW_PROP_FN

OperatorConstraints::flow_prop_fn for join and cross_join.

LATTICE_FOLD_REDUCE_FLOW_PROP_FN

OperatorConstraints::flow_prop_fn for lattice_fold and lattice_reduce.

NULL_WRITE_FN

OperatorConstraints::write_fn for the null operator - an operator that ignores all inputs and produces no output.

OPERATORS

All Hydroflow operators.

RANGE_0

Convenience range: exactly zero.

RANGE_1

Convenience range: exactly one.

RANGE_ANY

Convenience range: zero or more (any number).

Traits

RangeTrait

An object-safe version of RangeBounds.

Functions

find_node_op_constraints

Find an operator by GraphNode.

find_op_op_constraints

Find an operator by an AST Operator.

identity_write_iterator_fn

Helper to write the write_iterator portion of OperatorConstraints::write_fn output for unary identity operators.

null_write_iterator_fn

Helper to write the write_iterator portion of OperatorConstraints::write_fn output for the null operator - an operator that ignores all inputs and produces no output.

operator_lookup

Get the operator lookup table, generating it if needed.

Type Aliases

FlowPropFn

Type alias for OperatorConstraints::flow_prop_fn’s type.

WriteFn

Type alias for OperatorConstraints::write_fn’s type.