Struct hydroflow_plus::lang::graph::ops::WriteContextArgs

pub struct WriteContextArgs<'a> {

    pub root: &'a TokenStream,
    pub context: &'a Ident,
    pub hydroflow: &'a Ident,
    pub subgraph_id: GraphSubgraphId,
    pub node_id: GraphNodeId,
    pub op_span: Span,
    pub ident: &'a Ident,
    pub is_pull: bool,
    pub inputs: &'a [Ident],
    pub outputs: &'a [Ident],
    pub singleton_output_ident: &'a Ident,
    pub op_name: &'static str,
    pub op_inst: &'a OperatorInstance,
    pub arguments: &'a Punctuated<Expr, Comma>,
    pub arguments_handles: &'a Punctuated<Expr, Comma>,

}

Context arguments provided to OperatorConstraints::write_fn.

Fields
root: &'a TokenStream

hydroflow crate name for use #root::something.

context: &'a Ident

context ident, the name of the provided hydroflow::scheduled::Context.

hydroflow: &'a Ident

df ident, the name of the hydroflow::scheduled::graph::Hydroflow instance.

subgraph_id: GraphSubgraphId

Subgraph ID in which this operator is contained.

node_id: GraphNodeId

Node ID identifying this operator in the flat or partitioned graph meta-datastructure.

op_span: Span

The source span of this operator.

ident: &'a Ident

Ident the iterator or pullerator should be assigned to.

is_pull: bool

If a pull iterator (true) or pusherator (false) should be used.

inputs: &'a [Ident]

Input operator idents (or ref idents; used for pull).

outputs: &'a [Ident]

Output operator idents (or ref idents; used for push).

singleton_output_ident: &'a Ident

Ident for the singleton output of this operator, if any.

op_name: &'static str

Operator name.

op_inst: &'a OperatorInstance

Operator instance arguments object.

arguments: &'a Punctuated<Expr, Comma>

Arguments provided by the user into the operator as arguments. I.e. the a, b, c in -> my_op(a, b, c) -> .

These arguments include singleton postprocessing codegen, with std::cell::RefCell::borrow_mut code pre-generated.

arguments_handles: &'a Punctuated<Expr, Comma>

Same as Self::arguments but with only StateHandles, no borrowing code.

Implementations

impl WriteContextArgs<'_>

pub fn make_ident(&self, suffix: impl AsRef<str>) -> Ident

Generate a (almost certainly) unique identifier with the given suffix.

Includes the subgraph and node IDs in the generated identifier.

This will always return the same identifier for a given suffix.