Module hydroflow_plus::lang::process_singletons

Utility methods for processing singleton references: #my_var.

Functions

postprocess_singletons

Replaces singleton references #my_var with the code needed to actually get the value inside.

postprocess_singletons_handles

Same as postprocess_singletons but generates just the StateHandle ident rather than full RefCell borrowing code.

preprocess_singletons

Finds all the singleton references #my_var and appends them to found_idents. Returns the TokenStream but with the hashes removed from the varnames.