Module hydroflow_plus::lattices::algebra

Module for definiting algebraic structures and properties.

Functions

abelian_group

Defines an abelian group structure. An abelian group is a group where the operation f is commutative.

absorbing_element

Defines the absorbing_element property. An element z is absorbing if az = z and za = z for all a.

associativity

Defines the associativity property. a(bc) = (ab)c

commutative_monoid

Defines a commutative monoid structure. A commutative monoid is a monoid where the operation f is commutative.

commutative_ring

Defines a commutative ring structure. A commutative ring is a ring where the multiplication operation g is commutative.

commutativity

Defines the commutativity property. xy = yx

distributive

Defines the distributive property a(b+c) = ab + ac and (b+c)a = ba + ca

field

Defines a field structure. A field is a commutative ring where every element has a multiplicative inverse.

group

Defines a group structure. A group is a set of items along with an associative binary operation f an identity element zero and every element has an inverse element with respect to f

idempotency

Defines the idempotency property. xx = x

identity

Defines the identity property. An element e is the identity of f if ae = a and ea = a for all a.

inverse

Defines the inverse property. An element b is the inverse of a if ab = e and ba = e for some identity element e.

left_distributes

Defines the left distributive property a(b+c) = ab + ac

monoid

Defines a monoid structure. A monoid is a set of items along with an associative binary operation f and an identity element zero. The f operation combines two items and the zero element acts as the identity for f.

nonzero_inverse

Defines the non_zero inverse property. Every element except zero must have an inverse.

right_distributes

Defines the right distributive property. (b+c)a = ba + ca

ring

Defines a ring structure. A ring is a semiring with an inverse operation on f.

semigroup

Defines a semigroup structure. A semigroup is a set of items along with an associative binary operation f.

semiring

Defines a semiring structure. A semiring is a set of items along with two associative binary operations f and g, and two identity elements zero and one. f must be commutative and g must distribute over f. the zero of f must also be absorbing with respect to g.