Trait hydroflow_plus::lattices::cc_traits::BackMut

pub trait BackMut: CollectionMut {
    // Required method
    fn back_mut(&mut self) -> Option<Self::ItemMut<'_>>;
}

Collection exposing a mutable reference to its back element.

Required Methods

fn back_mut(&mut self) -> Option<Self::ItemMut<'_>>

Get a mutable reference to the back element of the collection.

Object Safety

This trait is not object safe.