Trait hydroflow_plus::lattices::cc_traits::CollectionMut

pub trait CollectionMut: Collection {
    type ItemMut<'a>: DerefMut<Target = Self::Item>
       where Self: 'a;

    // Required method
    fn upcast_item_mut<'short, 'long>(
        r: Self::ItemMut<'long>,
    ) -> Self::ItemMut<'short>
       where 'long: 'short,
             Self: 'long;
}

Abstract collection that can be mutably referenced.

Required Associated Types

type ItemMut<'a>: DerefMut<Target = Self::Item> where Self: 'a

Type of mutable references to items of the collection.

Required Methods

fn upcast_item_mut<'short, 'long>( r: Self::ItemMut<'long>, ) -> Self::ItemMut<'short>
where 'long: 'short, Self: 'long,

Changes an item mutable reference into a shorter lived mutable reference.

See the CollectionRef::upcast_item_ref function for more information. You can use the covariant_item_mut! macro to automatically implement this function.

Object Safety

This trait is not object safe.