Trait hydroflow_plus::lattices::cc_traits::Deque

pub trait Deque<T>: Stack<T> + Front { }

Immutable double-ended queue.

A double-ended queue (abbreviated to deque) is a generalization of a stack in which elements can be added and removed from both ends. Such a data structure provides two additional operations compared to regular stacks:

PushFront::push_front, which adds an element to the front of collection, and

PopFront::pop_front, which removes the front element of the collection.

This trait alias describes the immutables operations available on deques.