Trait hydroflow_plus::lattices::cc_traits::PopFront

pub trait PopFront: Collection {
    // Required method
    fn pop_front(&mut self) -> Option<Self::Item>;
}

Mutable collection where elements can be popped from the front.

Required Methods

fn pop_front(&mut self) -> Option<Self::Item>

Remove the front element of the collection and return it (if any).