Trait hydroflow_plus::lattices::cc_traits::SimpleCollectionMut

pub trait SimpleCollectionMut: CollectionMut {
    // Required method
    fn into_mut<'r>(r: Self::ItemMut<'r>) -> &'r mut Self::Item
       where Self: 'r;
}

Collection where each item reference can be converted into a standard “simple” rust reference.

This trait is particularly useful to avoid having to include where bounds of the form for<'r> T::ItemMut<'r>: Into<&'r mut T::Item>, which can currently lead the compiler to try to prove T: 'static (see https://github.com/rust-lang/rust/pull/96709#issuecomment-1182403490) for more details.

Required Methods

fn into_mut<'r>(r: Self::ItemMut<'r>) -> &'r mut Self::Item
where Self: 'r,

Object Safety

This trait is not object safe.