Trait hydroflow_plus::lattices::cc_traits::Stack

pub trait Stack<T>:
    Collection<Item = T>
    + Len
    + Back { }

Immutable stack data structure.

A stack provides two main operations:

PushBack::push_back, which adds an element to the collection, and

PopBack::pop_back, which removes the most recently added element that was not yet removed.

This trait alias describes the immutables operations derived from the two main operation above:

Len::len, returning the number of elements in the stack, and

Back::back, returning a reference to the most recently added element that was not yet removed.

Object Safety

This trait is not object safe.