Trait hydroflow_plus::lattices::cc_traits::WithCapacity

pub trait WithCapacity {
    // Required method
    fn with_capacity(capacity: usize) -> Self;
}

Collection that can be created with a minimum given capacity.

Required Methods

fn with_capacity(capacity: usize) -> Self

Creates a new instance of Self with the given minimum capacity.

Object Safety

This trait is not object safe.