Trait hydroflow_plus::lattices::generalized_hash_trie::GeneralizedHashTrie

pub trait GeneralizedHashTrie: Default + Sealed {
    type Row: VariadicExt + AsRefVariadicPartialEq;
    type Key: Eq + Hash;
    type Get: GeneralizedHashTrie;

    // Required methods
    fn insert(&mut self, row: Self::Row) -> bool;
    fn contains(&self, row: <Self::Row as VariadicExt>::AsRefVar<'_>) -> bool;
    fn iter(&self) -> impl Iterator<Item = &Self::Key>;
    fn get(&self, key: &Self::Key) -> Option<&Self::Get>;
    fn recursive_iter(
        &self
    ) -> impl Iterator<Item = <Self::Row as VariadicExt>::AsRefVar<'_>>;
}

node of a HashTrie

Required Associated Types

type Row: VariadicExt + AsRefVariadicPartialEq

Row variadic.

type Key: Eq + Hash

The first item in Row.

type Get: GeneralizedHashTrie

Type returned by Self::get.

Required Methods

fn insert(&mut self, row: Self::Row) -> bool

Inserts items into the hash trie.

fn contains(&self, row: <Self::Row as VariadicExt>::AsRefVar<'_>) -> bool

Returns true if the row is found, false otherwise.

fn iter(&self) -> impl Iterator<Item = &Self::Key>

Iterator for the keys (inner nodes) or elements (leaf nodes).

fn get(&self, key: &Self::Key) -> Option<&Self::Get>

Retrieves the value associated with the given key. returns an Option containing a reference to the value if found, or None if not found.

fn recursive_iter( &self ) -> impl Iterator<Item = <Self::Row as VariadicExt>::AsRefVar<'_>>

Iterate the sub-tree.

Object Safety

This trait is not object safe.