Trait hydroflow_plus::lattices::ght::GhtPrefixIter

pub trait GhtPrefixIter<KeyPrefix> {
    type Item: VariadicExt;

    // Required method
    fn prefix_iter<'a>(
        &'a self,
        prefix: KeyPrefix,
    ) -> impl Iterator<Item = <Self::Item as VariadicExt>::AsRefVar<'a>>
       where Self::Item: 'a;
}

A trait to iterate through the items in a Ght based on a prefix of the schema.

Required Associated Types

type Item: VariadicExt

the schema output

Required Methods

fn prefix_iter<'a>( &'a self, prefix: KeyPrefix, ) -> impl Iterator<Item = <Self::Item as VariadicExt>::AsRefVar<'a>>
where Self::Item: 'a,

given a prefix, return an iterator through the items below

Object Safety

This trait is not object safe.