Trait hydroflow_plus::lattices::ght_lazy::ColtNode

pub trait ColtNode: Sealed {
    type Schema: VariadicExt + Eq + Hash + Clone;
    type Storage: VariadicCollection;
    type SuffixSchema: VariadicExt + Eq + Hash + Clone;
    type Head: Eq + Hash;
    type Get;

    // Required method
    fn get(self, head: &Self::Head) -> Self::Get;
}

Virtual COLT “node” with an API similar to GeneralizedHashTrieNode.

The “current node” at depth d from the root in the Wang/Willsey/Suciu paper sits above an unbalanced trie. We emulate that here via a variadic of nodes at depth d in all the tries in the forest (that are height d or taller.) This is basically a GhtForest of subtries.

Required Associated Types

type Schema: VariadicExt + Eq + Hash + Clone

Schema variadic: the schema of the relation stored in this COLT. This type is the same in all Tries and nodes of the COLT.

type Storage: VariadicCollection

The type of Storage This type is the same in all Tries and nodes of the COLT

type SuffixSchema: VariadicExt + Eq + Hash + Clone

SuffixSchema variadic: the suffix of the schema from this node of the trie downward. The first entry in this variadic is of type Head. This type is the same in all Tries of the COLT (but changes as we traverse downward)

type Head: Eq + Hash

The type of the first column in the SuffixSchema This type is the same in all Tries of the COLT (but changes as we traverse downward)

type Get

Type returned by Self::get.

Required Methods

fn get(self, head: &Self::Head) -> Self::Get

On an Inner node, retrieves the value (child) associated with the given “head” key. returns an Option containing a reference to the value if found, or None if not found. On a Leaf node, returns None.