Trait hydroflow_plus::lattices::ght_lazy::ColumnLazyTrieNode

pub trait ColumnLazyTrieNode: GeneralizedHashTrieNode + Sealed {
    type Force: GeneralizedHashTrieNode;

    // Required methods
    fn into_iter(self) -> Option<impl Iterator<Item = Self::Schema>>;
    fn drain(&mut self) -> Option<impl Iterator<Item = Self::Schema>>;
    fn force(self) -> Option<Self::Force>;
    fn force_drain(&mut self) -> Option<Self::Force>;
}

COLT from Wang/Willsey/Suciu

In the paper, the COLT is an unbalanced trie that “grows upward” from leaves via the force method. Unbalanced tries don’t interact well with our Rust types, which want a node’s type to be defined via the type of its children, recursively – that means all paths need to be the same length!

To work around this, our COLT is a forest of balanced GHTs of increasing height, from 0 to the number of columns in the key - 1. Instead of the force method lengthening the path above a leaf by 1, it instead takes the leaf from the current trie and merges it in to the trie to the right which is 1 taller. A trait for the special behavior we need from the GHTs in a COLT forest#[derive(Debug, Clone)]

Required Associated Types

type Force: GeneralizedHashTrieNode

result of forceing a node

Required Methods

fn into_iter(self) -> Option<impl Iterator<Item = Self::Schema>>

into_iter for leaf elements, needed by force below

fn drain(&mut self) -> Option<impl Iterator<Item = Self::Schema>>

pull all the data out of this trie node but retain the reference

fn force(self) -> Option<Self::Force>

Force the generation of a parent node, as in the Wang/Willsey/Suciu COLT structure

fn force_drain(&mut self) -> Option<Self::Force>

Force the generation of a parent node but retain ref to this node

Object Safety

This trait is not object safe.