Trait hydroflow_plus::lattices::ght_lazy::GhtForest

pub trait GhtForest<SearchKey>: Sealed<SearchKey>
where
    SearchKey: VariadicExt,
{
    // Required method
    fn find_and_force(&mut self, search_key: SearchKey) -> bool;
}

A GhtForest trait with a force method that does the forcing+merging logic This trait will be recursive on the variadic of Ghts.

Required Methods

fn find_and_force(&mut self, search_key: SearchKey) -> bool

tries to find a trie with a matching leaf. if it finds such a trie, and the search_key is longer than the height, it will force the leaf into the next trie over and recurse. returns true if it forces (1 or more times), and false otherwise.