Struct hydroflow_plus::lattices::Min

pub struct Min<T>(/* private fields */);

A totally ordered min lattice. Merging returns the smaller value.

This means the lattice order is the reverse of what you might naturally expect: 0 is greater than 1.

Note that the Default::default() value for numeric type is MAX, not zero.

Implementations

impl<T> Min<T>

pub fn new(val: T) -> Min<T>

Create a new Min lattice instance from a T.

pub fn new_from(val: impl Into<T>) -> Min<T>

Create a new Min lattice instance from an Into<T> value.

pub fn as_reveal_ref(&self) -> &T

Reveal the inner value as a shared reference.

pub fn as_reveal_mut(&mut self) -> &mut T

Reveal the inner value as an exclusive reference.

pub fn into_reveal(self) -> T

Gets the inner by value, consuming self.