Trait hydroflow_plus::location::Deploy

pub trait Deploy<'a> {
    type ClusterId: Clone + 'static;
    type Process: Location<Meta = Self::Meta, Port = Self::ProcessPort> + HfSendOneToOne<Self::Process> + HfSendOneToMany<Self::Cluster, Self::ClusterId> + Clone;
    type Cluster: Location<Meta = Self::Meta, Port = Self::ClusterPort> + HfSendManyToOne<Self::Process, Self::ClusterId> + HfSendManyToMany<Self::Cluster, Self::ClusterId> + Cluster<'a, Id = Self::ClusterId> + Clone;
    type ProcessPort;
    type ClusterPort;
    type Meta: Default;
    type GraphId;
}

Required Associated Types

type ClusterId: Clone + 'static

Type of ID used to identify individual members of a cluster.

type Process: Location<Meta = Self::Meta, Port = Self::ProcessPort> + HfSendOneToOne<Self::Process> + HfSendOneToMany<Self::Cluster, Self::ClusterId> + Clone

type Cluster: Location<Meta = Self::Meta, Port = Self::ClusterPort> + HfSendManyToOne<Self::Process, Self::ClusterId> + HfSendManyToMany<Self::Cluster, Self::ClusterId> + Cluster<'a, Id = Self::ClusterId> + Clone

type ProcessPort

type ClusterPort

type Meta: Default

type GraphId

Type of ID used to switch between different subgraphs at runtime.

Object Safety

This trait is not object safe.