Trait hydroflow_plus::location::Location

pub trait Location<'a>: Clone {
    type Port;
    type Meta;

    // Required methods
    fn id(&self) -> usize;
    fn ir_leaves(&self) -> &'a RefCell<Vec<HfPlusLeaf>>;
    fn cycle_counter(&self) -> &RefCell<usize>;
    fn next_port(&self) -> Self::Port;
    fn update_meta(&mut self, meta: &Self::Meta);

    // Provided methods
    fn spin(&self) -> Stream<'a, (), Async, Self> { ... }
    fn spin_batch(
        &self,
        batch_size: impl Quoted<'a, usize> + Copy + 'a
    ) -> Stream<'a, (), Windowed, Self> { ... }
    fn source_stream<T, E: FuturesStream<Item = T> + Unpin>(
        &self,
        e: impl Quoted<'a, E>
    ) -> Stream<'a, T, Async, Self> { ... }
    fn source_external(&self) -> (Self::Port, Stream<'a, Bytes, Async, Self>)
       where Self: HfSendOneToOne<'a, Self> { ... }
    fn many_source_external<S, Cid>(
        &self
    ) -> (Self::Port, Stream<'a, Bytes, Async, Self>)
       where S: HfSendOneToMany<'a, Self, Cid> + Location<'a> { ... }
    fn source_iter<T, E: IntoIterator<Item = T>>(
        &self,
        e: impl Quoted<'a, E>
    ) -> Stream<'a, T, Windowed, Self> { ... }
    fn source_interval(
        &self,
        interval: impl Quoted<'a, Duration> + Copy + 'a
    ) -> Stream<'a, Instant, Async, Self> { ... }
    fn cycle<T, W>(&self) -> (HfCycle<'a, T, W, Self>, Stream<'a, T, W, Self>) { ... }

}

Required Associated Types

type Port

type Meta

Required Methods

fn id(&self) -> usize

fn ir_leaves(&self) -> &'a RefCell<Vec<HfPlusLeaf>>

A handle to the global list of IR leaves, which are the outputs of the program.

fn cycle_counter(&self) -> &RefCell<usize>

A handle to a counter of cycles within this location.

fn next_port(&self) -> Self::Port

fn update_meta(&mut self, meta: &Self::Meta)

Provided Methods

fn spin(&self) -> Stream<'a, (), Async, Self>

fn spin_batch( &self, batch_size: impl Quoted<'a, usize> + Copy + 'a ) -> Stream<'a, (), Windowed, Self>

fn source_stream<T, E: FuturesStream<Item = T> + Unpin>( &self, e: impl Quoted<'a, E> ) -> Stream<'a, T, Async, Self>

fn source_external(&self) -> (Self::Port, Stream<'a, Bytes, Async, Self>)
where Self: HfSendOneToOne<'a, Self>,

fn many_source_external<S, Cid>( &self ) -> (Self::Port, Stream<'a, Bytes, Async, Self>)
where S: HfSendOneToMany<'a, Self, Cid> + Location<'a>,

fn source_iter<T, E: IntoIterator<Item = T>>( &self, e: impl Quoted<'a, E> ) -> Stream<'a, T, Windowed, Self>

fn source_interval( &self, interval: impl Quoted<'a, Duration> + Copy + 'a ) -> Stream<'a, Instant, Async, Self>

fn cycle<T, W>(&self) -> (HfCycle<'a, T, W, Self>, Stream<'a, T, W, Self>)

Object Safety

This trait is not object safe.