Crate hydroflow_plus::pusherator

Pusherator generics and argument order conventions:

Next (being the next owned pusherator) should come first in generic arguments.

However next: Next in new(...) arguments should come last. This is so the rest of the arguments appear in the order data flows in.

Any closures Func should come before their arguments, so: <Func: Fn(A) -> B, A, B>

Modules

demux

filter

filter_map

flatten

for_each

inspect

map

null

partition

pivot

switch

tee

unzip

Structs

InputBuild

Traits

IteratorToPusherator

Pusherator

PusheratorBuild