Trait hydroflow_plus::scheduled::handoff::handoff_list::PortListSplit

pub trait PortListSplit<S, A>: PortList<S> + Sealed<S, A>
where
    S: Polarity,
    A: PortList<S>,
{
    type Suffix: PortList<S>;

    // Required method
    fn split_ctx(
        ctx: Self::Ctx<'_>,
    ) -> (<A as PortList<S>>::Ctx<'_>, <Self::Suffix as PortList<S>>::Ctx<'_>);
}

Trait for splitting a list of ports into two.

Required Associated Types

type Suffix: PortList<S>

The suffix, second half of the split.

Required Methods

fn split_ctx( ctx: Self::Ctx<'_>, ) -> (<A as PortList<S>>::Ctx<'_>, <Self::Suffix as PortList<S>>::Ctx<'_>)

Split the port list, returning the prefix and Self::Suffix as the two halves.

Object Safety

This trait is not object safe.