Trait hydroflow_plus::scheduled::handoff::HandoffMeta

pub trait HandoffMeta: Any {
    // Required methods
    fn any_ref(&self) -> &(dyn Any + 'static);
    fn is_bottom(&self) -> bool;
}

A handle onto the metadata part of a Handoff, with no element type.

Required Methods

fn any_ref(&self) -> &(dyn Any + 'static)

Helper to cast an instance of HandoffMeta to Any. In general you cannot cast between traits, including Any, but this helper method works around that limitation.

For implementors: the body of this method will generally just be { self }.

fn is_bottom(&self) -> bool

Return if the handoff is empty.