Module hydroflow_plus::scheduled::port

Organizational module for Hydroflow Send/RecvCtx structs and Input/OutputPort structs.

Structs

Port

Lightweight ID struct representing an input or output port for a Handoff added to a Hydroflow instance..

PortCtx

Wrapper around a handoff to differentiate between output and input.

Enums

RECV

An uninstantiable type used to tag port Polarity as receive.

SEND

An uninstantiable type used to tag port Polarity as send.

Traits

Polarity

An empty trait used to denote Polarity: either send or receive.

Type Aliases

RecvCtx

Recv-specific PortCtx. Input to receive from a handoff.

RecvPort

Recv-specific variant of Port. An input port.

SendCtx

Send-specific PortCtx. Output to send into a handoff.

SendPort

Send-specific variant of Port. An output port.