Struct hydroflow_plus::singleton::Singleton

pub struct Singleton<T, L, B> { /* private fields */ }

Implementations

impl<'a, T, L: Location<'a>, B> Singleton<T, L, B>

pub fn map<U, F: Fn(T) -> U + 'a>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Singleton<U, L, B>

pub fn flat_map_ordered<U, I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Stream<U, L, B>

pub fn flat_map_unordered<U, I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Stream<U, L, B>

pub fn filter<F: Fn(&T) -> bool + 'a>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Optional<T, L, B>

pub fn filter_map<U, F: Fn(T) -> Option<U> + 'a>( self, f: impl IntoQuotedMut<'a, F, L>, ) -> Optional<U, L, B>

pub fn zip<Other>(self, other: Other) -> <Self as ZipResult<'a, Other>>::Out
where Self: ZipResult<'a, Other, Location = L>,

pub fn continue_if<U>( self, signal: Optional<U, L, Bounded>, ) -> Optional<T, L, Bounded>
where Self: ZipResult<'a, Optional<(), L, Bounded>, Location = L, Out = Optional<(T, ()), L, Bounded>>,

pub fn continue_unless<U>( self, other: Optional<U, L, Bounded>, ) -> Optional<T, L, Bounded>
where Singleton<T, L, B>: ZipResult<'a, Optional<(), L, Bounded>, Location = L, Out = Optional<(T, ()), L, Bounded>>,

impl<'a, T, L: Location<'a> + NoTick, B> Singleton<T, Timestamped<L>, B>

pub unsafe fn latest_tick(self) -> Singleton<T, Tick<L>, Bounded>

Given a tick, returns a singleton value corresponding to a snapshot of the singleton as of that tick. The snapshot at tick t + 1 is guaranteed to include at least all relevant data that contributed to the snapshot at tick t.

Safety

Because this picks a snapshot of a singleton whose value is continuously changing, the output singleton has a non-deterministic value since the snapshot can be at an arbitrary point in time.

pub fn drop_timestamp(self) -> Optional<T, L, B>

impl<'a, T, L: Location<'a> + NoTick, B> Singleton<T, L, B>

pub fn timestamped(self, tick: &Tick<L>) -> Singleton<T, Timestamped<L>, B>

pub unsafe fn sample_eager(self) -> Stream<T, L, Unbounded>

Eagerly samples the singleton as fast as possible, returning a stream of snapshots with order corresponding to increasing prefixes of data contributing to the singleton.

Safety

At runtime, the singleton will be arbitrarily sampled as fast as possible, but due to non-deterministic batching and arrival of inputs, the output stream is non-deterministic.

pub unsafe fn sample_every( self, interval: impl QuotedWithContext<'a, Duration, L> + Copy + 'a, ) -> Stream<T, L, Unbounded>
where L: NoTimestamp,

Given a time interval, returns a stream corresponding to snapshots of the singleton value taken at various points in time. Because the input singleton may be Unbounded, there are no guarantees on what these snapshots are other than they represent the value of the singleton given some prefix of the streams leading up to it.

Safety

The output stream is non-deterministic in which elements are sampled, since this is controlled by a clock.

impl<'a, T, L: Location<'a>> Singleton<T, Tick<L>, Bounded>

pub fn all_ticks(self) -> Stream<T, Timestamped<L>, Unbounded>

pub fn latest(self) -> Singleton<T, Timestamped<L>, Unbounded>

pub fn defer_tick(self) -> Singleton<T, Tick<L>, Bounded>

pub fn persist(self) -> Stream<T, Tick<L>, Bounded>

pub fn delta(self) -> Optional<T, Tick<L>, Bounded>

pub fn into_stream(self) -> Stream<T, Tick<L>, Bounded>