Enum hydroflow_plus::tokio_util::either::Either

pub enum Either<L, R> {
    Left(L),
    Right(R),
}

Combines two different futures, streams, or sinks having the same associated types into a single type.

This type implements common asynchronous traits such as Future and those in Tokio.

Example

The following code will not work:

#[tokio::main]
async fn main() {
    let result = if some_condition() {
        some_async_function()
    } else {
        other_async_function() // <- Will print: "`if` and `else` have incompatible types"
    };

    println!("Result is {}", result.await);
}

When the output type is the same, we can wrap each future in Either to avoid the issue:

use tokio_util::either::Either;

#[tokio::main]
async fn main() {
    let result = if some_condition() {
        Either::Left(some_async_function())
    } else {
        Either::Right(other_async_function())
    };

    let value = result.await;
    println!("Result is {}", value);
}

Variants

Left(L)

Right(R)

Implementations

impl<L, R> Either<L, R>
where L: Listener, R: Listener,

pub async fn accept( &mut self ) -> Result<Either<(<L as Listener>::Io, <L as Listener>::Addr), (<R as Listener>::Io, <R as Listener>::Addr)>, Error>

Accepts a new incoming connection from this listener.

pub fn local_addr( &self ) -> Result<Either<<L as Listener>::Addr, <R as Listener>::Addr>, Error>

Returns the local address that this listener is bound to.