Struct hydroflow_plus::tokio_util::io::SyncIoBridge

pub struct SyncIoBridge<T> { /* private fields */ }

Use a tokio::io::AsyncRead synchronously as a std::io::Read or a tokio::io::AsyncWrite as a std::io::Write.

Implementations

impl<T> SyncIoBridge<T>
where T: AsyncWrite,

pub fn is_write_vectored(&self) -> bool

Determines if the underlying tokio::io::AsyncWrite target supports efficient vectored writes.

See tokio::io::AsyncWrite::is_write_vectored.

impl<T> SyncIoBridge<T>
where T: AsyncWrite + Unpin,

pub fn shutdown(&mut self) -> Result<(), Error>

Shutdown this writer. This method provides a way to call the AsyncWriteExt::shutdown function of the inner tokio::io::AsyncWrite instance.

Errors

This method returns the same errors as AsyncWriteExt::shutdown.

impl<T> SyncIoBridge<T>
where T: Unpin,

pub fn new(src: T) -> SyncIoBridge<T>

Use a tokio::io::AsyncRead synchronously as a std::io::Read or a tokio::io::AsyncWrite as a std::io::Write.

When this struct is created, it captures a handle to the current threadโs runtime with tokio::runtime::Handle::current. It is hence OK to move this struct into a separate thread outside the runtime, as created by e.g. tokio::task::spawn_blocking.

Stated even more strongly: to make use of this bridge, you must move it into a separate thread outside the runtime. The synchronous I/O will use the underlying handle to block on the backing asynchronous source, via tokio::runtime::Handle::block_on. As noted in the documentation for that function, an attempt to block_on from an asynchronous execution context will panic.

Wrapping !Unpin types

Use e.g. SyncIoBridge::new(Box::pin(src)).

Panics

This will panic if called outside the context of a Tokio runtime.

pub fn new_with_handle(src: T, rt: Handle) -> SyncIoBridge<T>

Use a tokio::io::AsyncRead synchronously as a std::io::Read or a tokio::io::AsyncWrite as a std::io::Write.

This is the same as SyncIoBridge::new, but allows passing an arbitrary handle and hence may be initially invoked outside of an asynchronous context.

pub fn into_inner(self) -> T

Consume this bridge, returning the underlying stream.