Trait hydroflow_plus::tokio_util::net::Listener

pub trait Listener {
    type Io: AsyncRead + AsyncWrite;
    type Addr;

    // Required methods
    fn poll_accept(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<(Self::Io, Self::Addr), Error>>;
    fn local_addr(&self) -> Result<Self::Addr, Error>;

    // Provided method
    fn accept(&mut self) -> ListenerAcceptFut<'_, Self> 
       where Self: Sized { ... }
}

A trait for a listener: TcpListener and UnixListener.

Required Associated Types

type Io: AsyncRead + AsyncWrite

The stream’s type of this listener.

type Addr

The socket address type of this listener.

Required Methods

fn poll_accept( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<(Self::Io, Self::Addr), Error>>

Polls to accept a new incoming connection to this listener.

fn local_addr(&self) -> Result<Self::Addr, Error>

Returns the local address that this listener is bound to.

Provided Methods

fn accept(&mut self) -> ListenerAcceptFut<'_, Self>
where Self: Sized,

Accepts a new incoming connection from this listener.