Module hydroflow_plus::tracing::metadata

Metadata describing trace data.

Structs

Kind

Indicates whether the callsite is a span or event.

Level

Describes the level of verbosity of a span or event.

LevelFilter

A filter comparable to a verbosity Level.

Metadata

Metadata describing a span or event.

ParseLevelError

Returned if parsing a Level fails.

ParseLevelFilterError

Indicates that a string could not be parsed to a valid level.