Module hydroflow_plus::util::deploy

Macros

launch

Structs

ConnectedDemux

ConnectedDirect

ConnectedTagged

DemuxDrain

DeployPorts

Contains runtime information passed by Hydro Deploy to a program, describing how to connect to other services and metadata about them.

MergeSource

TaggedSource

Enums

BoundConnection

RealizedServerPort

ServerBindConfig

ServerOrBound

ServerPort

Describes how to connect to a service which is listening on some port.

Traits

Connected

ConnectedSink

ConnectedSource

StreamSink

Functions

init

init_no_ack_start

launch_flow

Type Aliases

BufferedDrain

DynSink

DynStream

DynStreamSink

InitConfig