Function hydroflow_plus::util::bind_tcp

pub async fn bind_tcp<Item, Codec>(
    endpoint: SocketAddr,
    codec: Codec,
) -> Result<(Sender<(Item, SocketAddr)>, Receiver<Result<(<Codec as Decoder>::Item, SocketAddr), <Codec as Decoder>::Error>>, SocketAddr), Error>
where
    Item: 'static,
    Codec: 'static + Clone + Decoder + Encoder<Item>,
    <Codec as Encoder<Item>>::Error: Debug,

Create a listening tcp socket, and then as new connections come in, receive their data and forward it to a queue.