Trait hydroflow_plus::util::generalized_hash_trie::HashTrieNode

pub trait HashTrieNode: Default + Sealed {
    type Row: VariadicExt + AsRefVariadicPartialEq;
    type Key;
    type Get: HashTrieNode;

    // Required methods
    fn insert(&mut self, row: Self::Row);
    fn contains(&self, row: <Self::Row as VariadicExt>::AsRefVar<'_>) -> bool;
    fn keys(&self) -> impl Iterator<Item = &Self::Key>;
    fn get(&self, key: &Self::Key) -> Option<&Self::Get>;
    fn recursive_iter(
        &self
    ) -> impl Iterator<Item = <Self::Row as VariadicExt>::AsRefVar<'_>>;
}

node of a HashTrie

Required Associated Types

type Row: VariadicExt + AsRefVariadicPartialEq

Row variadic.

type Key

The first item in Row.

type Get: HashTrieNode

Type returned by Self::get.

Required Methods

fn insert(&mut self, row: Self::Row)

Inserts items into the hash trie.

fn contains(&self, row: <Self::Row as VariadicExt>::AsRefVar<'_>) -> bool

Returns true if the row is found, false otherwise.

fn keys(&self) -> impl Iterator<Item = &Self::Key>

Iterator for the keys.

fn get(&self, key: &Self::Key) -> Option<&Self::Get>

Retrieves the value associated with the given key. returns an Option containing a reference to the value if found, or None if not found.

fn recursive_iter( &self ) -> impl Iterator<Item = <Self::Row as VariadicExt>::AsRefVar<'_>>

Iterate the sub-tree.

Object Safety

This trait is not object safe.