Type Alias hydroflow_plus::util::TcpFramedSink

pub type TcpFramedSink<T> = Sender<(T, SocketAddr)>;

A framed TCP Sink (sending).

Aliased Type

struct TcpFramedSink<T> { /* private fields */ }

Implementations

impl<T> Sender<T>

pub async fn send(&self, item: T) -> Result<(), SendError<T>>

Asynchronously sends value to the receiver.

pub fn try_send(&self, item: T) -> Result<(), TrySendError<T>>

Tries to send the value to the receiver without blocking.

Returns an error if the destination is closed or if the buffer is at capacity.

TrySendError::Full will never be returned if this is an unbounded channel.

pub fn close_this_sender(&mut self)

Close this sender. No more messages can be sent from this sender.

Note that this only closes the channel from the view-point of this sender. The channel remains open until all senders have gone away, or until the Receiver closes the channel.

pub fn is_closed(&self) -> bool

If this sender or the corresponding Receiver is closed.