Trait hydroflow_plus::variadics::VariadicExt

pub trait VariadicExt: Variadic + Sealed {
    type Extend<Suffix: VariadicExt>: VariadicExt;
    type Reverse: VariadicExt;

    const LEN: usize;

    // Required methods
    fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>
       where Suffix: VariadicExt;
    fn reverse(self) -> Self::Reverse;
}

Extension methods/types for Variadics.

This is a sealed trait.

Required Associated Types

type Extend<Suffix: VariadicExt>: VariadicExt

Creates a new (longer) variadic type by appending Suffix onto the end of this variadc.

type Reverse: VariadicExt

The reverse of this variadic type.

Required Associated Constants

const LEN: usize

The number of items in this variadic (its length).

Required Methods

fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>where Suffix: VariadicExt,

Extends this variadic value by appending suffix onto the end.

fn reverse(self) -> Self::Reverse

Reverses this variadic value.