Module hydroflow_plus::variadics::variadic_collections

module of collection types for variadics

Structs

DuplicateCounted

Iterator helper for VariadicCountedHashSet::into_iter.

VariadicColumnMultiset

Column storage for Variadic tuples of type Schema An alternative to VariadicHashMultiset

VariadicCountedHashSet

HashMap keyed on Variadics of (owned value, count) pairs, allows for lookups with RefVariadics.

VariadicHashSet

HashSet that stores Variadics of owned values but allows for lookups with RefVariadics as well

Traits

VariadicCollection

Trait for a set of Variadic Tuples

VariadicMultiset

Trait for a multiset of Tuples

VariadicSet

trait for sets or multisets of variadics