Crate hydroflow_plus

Re-exports

pub use stream::Stream;

pub use location::Cluster;

pub use location::ClusterSpec;

pub use location::Deploy;

pub use location::LocalDeploy;

pub use location::Location;

pub use location::MultiGraph;

pub use location::ProcessSpec;

pub use location::SingleProcessGraph;

pub use cycle::HfCycle;

pub use builder::FlowBuilder;

Modules

bincode

Bincode is a crate for encoding and decoding using a tiny binary serialization strategy. Using it, you can easily go from having an object in memory, quickly serialize it to bytes, and then deserialize it back just as fast!

builder

bytes

Provides abstractions for working with bytes.

compiled

Hydroflow’s inner (intra-subgraph) compiled layer.

cycle

futures

Abstractions for asynchronous programming.

instant

itertools

Extra iterator adaptors, functions and macros.

lang

Hydroflow surface syntax

lattices

The lattices Crate

location

props

Hydroflow lattice and flow properties, very WIP.

pusherator

Pusherator generics and argument order conventions:

runtime_support

rustc_hash

Fast, non-cryptographic hash used by rustc and Firefox.

scheduled

Hydroflow’s outer scheduled layer. Deals with inter-subgraph runtime data-passing and scheduling.

serde

Serde

serde_json

Serde JSON

stream

tokio

A runtime for writing reliable network applications without compromising speed.

tokio_stream

Stream utilities for Tokio.

tokio_util

Utilities for working with Tokio.

util

Helper utilities for the Hydroflow surface syntax.

variadics

Variadics

Macros

assert_var_impl

Asserts that the variable’s type implements the given traits.

datalog

Generate a Hydroflow instance from Datalog code.

hydroflow_expect_warnings

Tests that the given warnings are emitted by the hydroflow macro invocation.

hydroflow_parser

Parse Hydroflow “surface syntax” without emitting code.

hydroflow_syntax

Create a Hydroflow instance using Hydroflow’s custom “surface syntax.”

hydroflow_syntax_noemit

hydroflow_syntax! but will not emit any diagnostics (errors, warnings, etc.).

monotonic_fn

Checks that the given closure is a monotonic function. For now does nothing.

morphism

Checks that the given closure is a morphism. For now does nothing.

rassert

assert! but returns a [Result<(), String>] instead of panicking.

rassert_eq

assert_eq! but returns a [Result<(), String>] instead of panicking.

var_args

Variadic patterns macro.

var_expr

Variadic expressions (values) macro.

var_type

Variadic types macro.

Structs

HfBuilt

Hydroflow

A Hydroflow graph. Owns, schedules, and runs the compiled subgraphs.

RuntimeContext

Enums

Never

Stand-in for the nightly “never” type !

Attribute Macros

main

test