Module hydroflow_plus::lattices::collections

Simple singleton or array collection with cc_traits implementations.

Structs

ArrayMap

An array wrapper representing a fixed-size map.

ArraySet

An array wrapper representing a fixed-size set (modulo duplicate items).

EmptyMap

A key-value entry wrapper representing a singleton map.

EmptySet

A type that will always be an empty set.

OptionMap

A key-value entry wrapper around Option<(K, V)> representing a singleton or empty map.

OptionSet

A wrapper around Option, representing either a singleton or empty set.

SingletonMap

A key-value entry wrapper representing a singleton map.

SingletonSet

A wrapper around an item, representing a singleton set.

VecMap

A Vec-wrapper representing a naively implemented map.

VecSet

A Vec-wrapper representing a naively-implemented set.

Traits

MapMapValues

Trait for transforming the values of a map without changing the overall type of the data structure.