Trait hydroflow_plus::lattices::collections::MapMapValues

pub trait MapMapValues<OldVal> {
    type MapValue<NewVal>;

    // Required method
    fn map_values<NewVal, MapFn>(self, map_fn: MapFn) -> Self::MapValue<NewVal>
       where MapFn: FnMut(OldVal) -> NewVal;
}

Trait for transforming the values of a map without changing the overall type of the data structure.

Required Associated Types

type MapValue<NewVal>

Output type, should be Self but with OldVal replaced with NewVal.

Required Methods

fn map_values<NewVal, MapFn>(self, map_fn: MapFn) -> Self::MapValue<NewVal>
where MapFn: FnMut(OldVal) -> NewVal,

Map the values into using the map_fn.

Object Safety

This trait is not object safe.