Trait hydroflow_plus::lattices::ght::GeneralizedHashTrieNode

pub trait GeneralizedHashTrieNode: Default + Sealed {
    type Schema: VariadicExt + Eq + Hash + Clone + SplitBySuffix<Self::SuffixSchema>;
    type KeyType: VariadicExt + Eq + Hash + Clone;
    type ValType: VariadicExt + Eq + Hash + Clone;
    type Storage: VariadicCollection<Schema = Self::Schema> + Default + IntoIterator<Item = Self::Schema>;
    type SuffixSchema: VariadicExt + Eq + Hash + Clone;
    type Head: Eq + Hash + Clone;
    type DeepJoin<Other: GeneralizedHashTrieNode, Storage>
       where (Self, Other): DeepJoinLatticeBimorphism<Storage>;

    // Required methods
    fn new_from(input: impl IntoIterator<Item = Self::Schema>) -> Self;
    fn merge_node(&mut self, other: Self) -> bool;
    fn static_height() -> usize;
    fn is_leaf(&self) -> bool;
    fn insert(&mut self, row: Self::Schema) -> bool;
    fn contains<'a>(
        &'a self,
        row: <Self::Schema as VariadicExt>::AsRefVar<'a>,
    ) -> bool;
    fn recursive_iter(
        &self,
    ) -> impl Iterator<Item = <Self::Schema as VariadicExt>::AsRefVar<'_>>;
    fn find_containing_leaf(
        &self,
        row: <Self::Schema as VariadicExt>::AsRefVar<'_>,
    ) -> Option<&GhtLeaf<Self::Schema, Self::ValType, Self::Storage>>;

    // Provided method
    fn height(&self) -> usize { ... }
}

GeneralizedHashTrieNode trait

Required Associated Types

type Schema: VariadicExt + Eq + Hash + Clone + SplitBySuffix<Self::SuffixSchema>

Schema variadic: the schema of the relation stored in this trie. This type is the same in all nodes of the trie.

type KeyType: VariadicExt + Eq + Hash + Clone

The prefix of columns in Schema that the trie is keyed on This type is the same in all nodes of the trie.

type ValType: VariadicExt + Eq + Hash + Clone

The suffix of columns in Schema that are not part of the trie keys This type is the same in all nodes of the trie.

type Storage: VariadicCollection<Schema = Self::Schema> + Default + IntoIterator<Item = Self::Schema>

The type that holds the data in the leaves

type SuffixSchema: VariadicExt + Eq + Hash + Clone

SuffixSchema variadic: the suffix of the schema from this node of the trie downward. The first entry in this variadic is of type Head.

type Head: Eq + Hash + Clone

Head is the first field in SuffixSchema

type DeepJoin<Other: GeneralizedHashTrieNode, Storage> where (Self, Other): DeepJoinLatticeBimorphism<Storage>

Bimorphism for joining on full tuple keys (all GhtInner keys) in the trie

Required Methods

fn new_from(input: impl IntoIterator<Item = Self::Schema>) -> Self

Create a new Ght from the iterator.

fn merge_node(&mut self, other: Self) -> bool

Merge a matching Ght node into this one

fn static_height() -> usize

Return the height of this node in the GhT. Leaf = 0.

fn is_leaf(&self) -> bool

report whether node is a leaf node; else an inner node

fn insert(&mut self, row: Self::Schema) -> bool

Inserts items into the hash trie.

fn contains<'a>( &'a self, row: <Self::Schema as VariadicExt>::AsRefVar<'a>, ) -> bool

Returns true if the (entire) row is found in the trie, false otherwise. See get() below to look just for “head” keys in this node

fn recursive_iter( &self, ) -> impl Iterator<Item = <Self::Schema as VariadicExt>::AsRefVar<'_>>

Iterate through the (entire) rows stored in this HashTrie. Returns Variadics, not tuples.

fn find_containing_leaf( &self, row: <Self::Schema as VariadicExt>::AsRefVar<'_>, ) -> Option<&GhtLeaf<Self::Schema, Self::ValType, Self::Storage>>

return the leaf in the trie that contains this row

Provided Methods

fn height(&self) -> usize

Report the height of the tree if its not empty. This is the length of a root to leaf path -1. E.g. if we have GhtInner<GhtInner<GhtLeaf…>> the height is 2

Object Safety

This trait is not object safe.