Struct hydroflow_plus::stream::Stream

pub struct Stream<'a, T, W, N: Location + Clone> { /* private fields */ }

An infinite stream of elements of type T.

Type Parameters:

'a: the lifetime of the final Hydroflow graph, which constraints which values can be captured in closures passed to operators

T: the type of elements in the stream

W: the windowing semantics of the stream, which is either Async or Windowed

N: the type of the node that the stream is materialized on

Implementations

impl<'a, T, W, N: Location + Clone> Stream<'a, T, W, N>

pub fn map<U, F: Fn(T) -> U + 'a>( self, f: impl IntoQuotedMut<'a, F> ) -> Stream<'a, U, W, N>

pub fn flat_map<U, I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a>( self, f: impl IntoQuotedMut<'a, F> ) -> Stream<'a, U, W, N>

pub fn enumerate(self) -> Stream<'a, (usize, T), W, N>

pub fn inspect<F: Fn(&T) + 'a>( self, f: impl IntoQuotedMut<'a, F> ) -> Stream<'a, T, W, N>

pub fn filter<F: Fn(&T) -> bool + 'a>( self, f: impl IntoQuotedMut<'a, F> ) -> Stream<'a, T, W, N>

pub fn filter_map<U, F: Fn(T) -> Option<U> + 'a>( self, f: impl IntoQuotedMut<'a, F> ) -> Stream<'a, U, W, N>

pub fn cross_product<O>( self, other: Stream<'a, O, W, N> ) -> Stream<'a, (T, O), W, N>

pub fn union(self, other: Stream<'a, T, W, N>) -> Stream<'a, T, W, N>

pub fn for_each<F: Fn(T) + 'a>(self, f: impl IntoQuotedMut<'a, F>)

pub fn dest_sink<S: Unpin + Sink<T> + 'a>(self, sink: impl Quoted<'a, S>)

pub fn all_ticks(self) -> Stream<'a, T, Windowed, N>

pub fn assume_windowed(self) -> Stream<'a, T, Windowed, N>

impl<'a, T, N: Location + Clone> Stream<'a, T, Async, N>

pub fn tick_batch(self) -> Stream<'a, T, Windowed, N>

impl<'a, T, N: Location + Clone> Stream<'a, T, Windowed, N>

pub fn fold<A, I: Fn() -> A + 'a, C: Fn(&mut A, T)>( self, init: impl IntoQuotedMut<'a, I>, comb: impl IntoQuotedMut<'a, C> ) -> Stream<'a, A, Windowed, N>

pub fn reduce<C: Fn(&mut T, T) + 'a>( self, comb: impl IntoQuotedMut<'a, C> ) -> Stream<'a, T, Windowed, N>

pub fn count(self) -> Stream<'a, usize, Windowed, N>

pub fn delta(self) -> Stream<'a, T, Windowed, N>

pub fn unique(self) -> Stream<'a, T, Windowed, N>
where T: Eq + Hash,

pub fn filter_not_in( self, other: Stream<'a, T, Windowed, N> ) -> Stream<'a, T, Windowed, N>
where T: Eq + Hash,

pub fn sample_every( self, duration: impl Quoted<'a, Duration> + Copy + 'a ) -> Stream<'a, T, Windowed, N>

impl<'a, T: Clone, W, N: Location + Clone> Stream<'a, &T, W, N>

pub fn cloned(self) -> Stream<'a, T, W, N>

impl<'a, K, V1, W, N: Location + Clone> Stream<'a, (K, V1), W, N>

pub fn join<W2, V2>( self, n: Stream<'a, (K, V2), W2, N> ) -> Stream<'a, (K, (V1, V2)), W, N>
where K: Eq + Hash,

pub fn anti_join<W2>(self, n: Stream<'a, K, W2, N>) -> Stream<'a, (K, V1), W, N>
where K: Eq + Hash,

impl<'a, K: Eq + Hash, V, N: Location + Clone> Stream<'a, (K, V), Windowed, N>

pub fn fold_keyed<A, I: Fn() -> A + 'a, C: Fn(&mut A, V) + 'a>( self, init: impl IntoQuotedMut<'a, I>, comb: impl IntoQuotedMut<'a, C> ) -> Stream<'a, (K, A), Windowed, N>

pub fn reduce_keyed<F: Fn(&mut V, V) + 'a>( self, comb: impl IntoQuotedMut<'a, F> ) -> Stream<'a, (K, V), Windowed, N>

impl<'a, T, W, N: Location + Clone> Stream<'a, T, W, N>

pub fn send_bincode<N2: Location + Clone, V, CoreType>( self, other: &N2 ) -> Stream<'a, N::Out<CoreType>, Async, N2>
where N: HfSend<N2, V, In<CoreType> = T>, CoreType: Serialize + DeserializeOwned,

pub fn send_bytes<N2: Location + Clone, V>( self, other: &N2 ) -> Stream<'a, N::Out<Bytes>, Async, N2>
where N: HfSend<N2, V, In<Bytes> = T>,

pub fn send_bincode_interleaved<N2: Location + Clone, Tag, CoreType, V>( self, other: &N2 ) -> Stream<'a, T, Async, N2>
where N: HfSend<N2, V, In<CoreType> = T, Out<CoreType> = (Tag, T)>, CoreType: Serialize + DeserializeOwned,

pub fn send_bytes_interleaved<N2: Location + Clone, Tag, V>( self, other: &N2 ) -> Stream<'a, Bytes, Async, N2>
where N: HfSend<N2, V, In<Bytes> = T, Out<Bytes> = (Tag, Bytes)>,

pub fn broadcast_bincode<N2: Location + Cluster<'a> + Clone, V>( self, other: &N2 ) -> Stream<'a, N::Out<T>, Async, N2>
where N: HfSend<N2, V, In<T> = (N2::Id, T)>, T: Serialize + DeserializeOwned, N2::Id: Clone,

pub fn broadcast_bincode_interleaved<N2: Location + Cluster<'a> + Clone, Tag, V>( self, other: &N2 ) -> Stream<'a, T, Async, N2>
where N: HfSend<N2, V, In<T> = (N2::Id, T), Out<T> = (Tag, T)>, T: Serialize + DeserializeOwned, N2::Id: Clone,

pub fn broadcast_bytes<N2: Location + Cluster<'a> + Clone, V>( self, other: &N2 ) -> Stream<'a, N::Out<Bytes>, Async, N2>
where N: HfSend<N2, V, In<Bytes> = (N2::Id, T)>, N2::Id: Clone,

pub fn broadcast_bytes_interleaved<N2: Location + Cluster<'a> + Clone, Tag, V>( self, other: &N2 ) -> Stream<'a, Bytes, Async, N2>
where N: HfSend<N2, V, In<Bytes> = (N2::Id, T), Out<Bytes> = (Tag, Bytes)>, N2::Id: Clone,