Struct hydroflow_plus::stream::Stream

pub struct Stream<T, W, C, N> { /* private fields */ }

An infinite stream of elements of type T.

Type Parameters:

'a: the lifetime of the final Hydroflow graph, which constraints which values can be captured in closures passed to operators

T: the type of elements in the stream

W: the boundedness of the stream, which is either Bounded or Unbounded

C: the tick domain of the stream, which is either Tick or NoTick

N: the type of the node that the stream is materialized on

Implementations

impl<'a, T, W, C, N: Location<'a>> Stream<T, W, C, N>

pub fn map<U, F: Fn(T) -> U + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<U, W, C, N>

pub fn flat_map<U, I: IntoIterator<Item = U>, F: Fn(T) -> I + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<U, W, C, N>

pub fn filter<F: Fn(&T) -> bool + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<T, W, C, N>

pub fn filter_map<U, F: Fn(T) -> Option<U> + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<U, W, C, N>

pub fn cross_singleton<O>( self, other: impl Into<Optional<O, Bounded, C, N>>, ) -> Stream<(T, O), W, C, N>
where O: Clone,

pub fn cross_product<O>( self, other: Stream<O, W, C, N>, ) -> Stream<(T, O), W, C, N>
where T: Clone, O: Clone,

pub fn union(self, other: Stream<T, W, C, N>) -> Stream<T, W, C, N>

pub fn unique(self) -> Stream<T, W, C, N>
where T: Eq + Hash,

pub fn dest_sink<S: Unpin + Sink<T> + 'a>(self, sink: impl Quoted<'a, S>)

impl<'a, T, N: Location<'a>> Stream<T, Bounded, Tick, N>

pub fn all_ticks(self) -> Stream<T, Unbounded, NoTick, N>

pub fn persist(self) -> Stream<T, Bounded, Tick, N>
where T: Clone,

pub fn defer_tick(self) -> Stream<T, Bounded, Tick, N>

pub fn inspect<F: Fn(&T) + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<T, Bounded, Tick, N>

pub fn first(self) -> Optional<T, Bounded, Tick, N>

pub fn continue_if<U>( self, signal: Optional<U, Bounded, Tick, N>, ) -> Stream<T, Bounded, Tick, N>

Allow this stream through if the other stream has elements, otherwise the output is empty.

pub fn continue_unless<U>( self, other: Optional<U, Bounded, Tick, N>, ) -> Stream<T, Bounded, Tick, N>

Allow this stream through if the other stream is empty, otherwise the output is empty.

pub fn enumerate(self) -> Stream<(usize, T), Bounded, Tick, N>

pub fn fold<A, I: Fn() -> A + 'a, F: Fn(&mut A, T)>( self, init: impl IntoQuotedMut<'a, I>, comb: impl IntoQuotedMut<'a, F>, ) -> Singleton<A, Bounded, Tick, N>

pub fn reduce<F: Fn(&mut T, T) + 'a>( self, comb: impl IntoQuotedMut<'a, F>, ) -> Optional<T, Bounded, Tick, N>

pub fn max(self) -> Optional<T, Bounded, Tick, N>
where T: Ord,

pub fn min(self) -> Optional<T, Bounded, Tick, N>
where T: Ord,

pub fn sort(self) -> Stream<T, Bounded, Tick, N>
where T: Ord,

pub fn count(self) -> Singleton<usize, Bounded, Tick, N>

pub fn delta(self) -> Stream<T, Bounded, Tick, N>

impl<'a, T, W, N: Location<'a>> Stream<T, W, NoTick, N>

pub fn tick_batch(self) -> Stream<T, Bounded, Tick, N>

pub fn tick_prefix(self) -> Stream<T, Bounded, Tick, N>
where T: Clone,

pub fn inspect<F: Fn(&T) + 'a>( self, f: impl IntoQuotedMut<'a, F>, ) -> Stream<T, W, NoTick, N>

impl<'a, T, W, N: Location<'a>> Stream<T, W, NoTick, N>

pub fn for_each<F: Fn(T) + 'a>(self, f: impl IntoQuotedMut<'a, F>)

impl<'a, T, N: Location<'a>> Stream<T, Unbounded, NoTick, N>

pub fn sample_every( self, duration: impl Quoted<'a, Duration> + Copy + 'a, ) -> Stream<T, Unbounded, NoTick, N>

pub fn fold<A, I: Fn() -> A + 'a, F: Fn(&mut A, T)>( self, init: impl IntoQuotedMut<'a, I>, comb: impl IntoQuotedMut<'a, F>, ) -> Singleton<A, Unbounded, NoTick, N>

pub fn reduce<F: Fn(&mut T, T) + 'a>( self, comb: impl IntoQuotedMut<'a, F>, ) -> Optional<T, Unbounded, NoTick, N>

pub fn max(self) -> Optional<T, Unbounded, NoTick, N>
where T: Ord,

pub fn min(self) -> Optional<T, Unbounded, NoTick, N>
where T: Ord,

impl<'a, T, C, N: Location<'a>> Stream<T, Bounded, C, N>

pub fn filter_not_in( self, other: Stream<T, Bounded, C, N>, ) -> Stream<T, Bounded, C, N>
where T: Eq + Hash,

impl<'a, T: Clone, W, C, N: Location<'a>> Stream<&T, W, C, N>

pub fn cloned(self) -> Stream<T, W, C, N>

impl<'a, K, V1, W, C, N: Location<'a>> Stream<(K, V1), W, C, N>

pub fn join<W2, V2>( self, n: Stream<(K, V2), W2, C, N>, ) -> Stream<(K, (V1, V2)), W, C, N>
where K: Eq + Hash,

pub fn anti_join<W2>(self, n: Stream<K, W2, C, N>) -> Stream<(K, V1), W, C, N>
where K: Eq + Hash,

impl<'a, K: Eq + Hash, V, N: Location<'a>> Stream<(K, V), Bounded, Tick, N>

pub fn fold_keyed<A, I: Fn() -> A + 'a, F: Fn(&mut A, V) + 'a>( self, init: impl IntoQuotedMut<'a, I>, comb: impl IntoQuotedMut<'a, F>, ) -> Stream<(K, A), Bounded, Tick, N>

pub fn reduce_keyed<F: Fn(&mut V, V) + 'a>( self, comb: impl IntoQuotedMut<'a, F>, ) -> Stream<(K, V), Bounded, Tick, N>

impl<'a, T, W, N: Location<'a>> Stream<T, W, NoTick, N>

pub fn decouple_process<P2>( self, other: &Process<'a, P2>, ) -> Stream<T, Unbounded, NoTick, Process<'a, P2>>
where N: CanSend<'a, Process<'a, P2>, In<T> = T, Out<T> = T>, T: Clone + Serialize + DeserializeOwned,

pub fn decouple_cluster<C2, Tag>( self, other: &Cluster<'a, C2>, ) -> Stream<T, Unbounded, NoTick, Cluster<'a, C2>>
where N: CanSend<'a, Cluster<'a, C2>, In<T> = (ClusterId<C2>, T), Out<T> = (Tag, T)>, T: Clone + Serialize + DeserializeOwned,

pub fn send_bincode<N2: Location<'a>, CoreType>( self, other: &N2, ) -> Stream<N::Out<CoreType>, Unbounded, NoTick, N2>
where N: CanSend<'a, N2, In<CoreType> = T>, CoreType: Serialize + DeserializeOwned,

pub fn send_bincode_external<N2: 'a, CoreType>( self, other: &ExternalProcess<'_, N2>, ) -> ExternalBincodeStream<N::Out<CoreType>>
where N: CanSend<'a, ExternalProcess<'a, N2>, In<CoreType> = T, Out<CoreType> = CoreType>, CoreType: Serialize + DeserializeOwned,

pub fn send_bytes<N2: Location<'a>>( self, other: &N2, ) -> Stream<N::Out<Bytes>, Unbounded, NoTick, N2>
where N: CanSend<'a, N2, In<Bytes> = T>,

pub fn send_bytes_external<N2: 'a>( self, other: &ExternalProcess<'_, N2>, ) -> ExternalBytesPort
where N: CanSend<'a, ExternalProcess<'a, N2>, In<Bytes> = T, Out<Bytes> = Bytes>,

pub fn send_bincode_interleaved<N2: Location<'a>, Tag, CoreType>( self, other: &N2, ) -> Stream<CoreType, Unbounded, NoTick, N2>
where N: CanSend<'a, N2, In<CoreType> = T, Out<CoreType> = (Tag, CoreType)>, CoreType: Serialize + DeserializeOwned,

pub fn send_bytes_interleaved<N2: Location<'a>, Tag>( self, other: &N2, ) -> Stream<Bytes, Unbounded, NoTick, N2>
where N: CanSend<'a, N2, In<Bytes> = T, Out<Bytes> = (Tag, Bytes)>,

pub fn broadcast_bincode<C2>( self, other: &Cluster<'a, C2>, ) -> Stream<N::Out<T>, Unbounded, NoTick, Cluster<'a, C2>>
where N: CanSend<'a, Cluster<'a, C2>, In<T> = (ClusterId<C2>, T)>, T: Clone + Serialize + DeserializeOwned,

pub fn broadcast_bincode_interleaved<C2, Tag>( self, other: &Cluster<'a, C2>, ) -> Stream<T, Unbounded, NoTick, Cluster<'a, C2>>
where N: CanSend<'a, Cluster<'a, C2>, In<T> = (ClusterId<C2>, T), Out<T> = (Tag, T)> + 'a, T: Clone + Serialize + DeserializeOwned,

pub fn broadcast_bytes<C2>( self, other: &Cluster<'a, C2>, ) -> Stream<N::Out<Bytes>, Unbounded, NoTick, Cluster<'a, C2>>
where N: CanSend<'a, Cluster<'a, C2>, In<Bytes> = (ClusterId<C2>, T)> + 'a, T: Clone,

pub fn broadcast_bytes_interleaved<C2, Tag>( self, other: &Cluster<'a, C2>, ) -> Stream<Bytes, Unbounded, NoTick, Cluster<'a, C2>>
where N: CanSend<'a, Cluster<'a, C2>, In<Bytes> = (ClusterId<C2>, T), Out<Bytes> = (Tag, Bytes)> + 'a, T: Clone,