Trait hydroflow_plus::variadics::VariadicExt

pub trait VariadicExt: Variadic + Sealed {
    type Extend<Suffix: VariadicExt>: VariadicExt;
    type Reverse: VariadicExt;
    type AsRefVar<'a>: RefVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>>
       where Self: 'a;
    type AsMutVar<'a>: MutVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>>
       where Self: 'a;
    type IterAnyRef<'a>: Iterator<Item = &'a (dyn Any + 'static)>
       where Self: 'static;
    type IterAnyMut<'a>: Iterator<Item = &'a mut (dyn Any + 'static)>
       where Self: 'static;
    type IntoOption;
    type IntoVec: VecVariadic<UnVec = Self> + Default;

    const LEN: usize;

    // Required methods
    fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>
       where Suffix: VariadicExt;
    fn reverse(self) -> Self::Reverse;
    fn reverse_ref(
        this: Self::AsRefVar<'_>,
    ) -> <Self::Reverse as VariadicExt>::AsRefVar<'_>;
    fn as_ref_var(&self) -> Self::AsRefVar<'_>;
    fn as_mut_var(&mut self) -> Self::AsMutVar<'_>;
    fn iter_any_ref(&self) -> Self::IterAnyRef<'_>
       where Self: 'static;
    fn iter_any_mut(&mut self) -> Self::IterAnyMut<'_>
       where Self: 'static;
    fn into_option(self) -> Self::IntoOption;
    fn into_singleton_vec(self) -> Self::IntoVec;

    // Provided methods
    fn len(&self) -> usize { ... }
    fn is_empty(&self) -> bool { ... }
}

Extension methods/types for Variadics.

This is a sealed trait.

Required Associated Types

type Extend<Suffix: VariadicExt>: VariadicExt

Creates a new (longer) variadic type by appending Suffix onto the end of this variadc.

type Reverse: VariadicExt

The reverse of this variadic type.

type AsRefVar<'a>: RefVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>> where Self: 'a

This as a variadic of references.

type AsMutVar<'a>: MutVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>> where Self: 'a

This as a variadic of exclusive (mut) references.

type IterAnyRef<'a>: Iterator<Item = &'a (dyn Any + 'static)> where Self: 'static

Iterator type returned by Self::iter_any_ref.

type IterAnyMut<'a>: Iterator<Item = &'a mut (dyn Any + 'static)> where Self: 'static

Iterator type returned by Self::iter_any_mut.

type IntoOption

type for all elements of the variadic being wrapped in Option

type IntoVec: VecVariadic<UnVec = Self> + Default

type for all elements of the variadic being wrapped in Vec

Required Associated Constants

const LEN: usize

The number of items in this variadic (its length).

Required Methods

fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>
where Suffix: VariadicExt,

Extends this variadic value by appending suffix onto the end.

fn reverse(self) -> Self::Reverse

Reverses this variadic value.

fn reverse_ref( this: Self::AsRefVar<'_>, ) -> <Self::Reverse as VariadicExt>::AsRefVar<'_>

Reverses an AsRefVar variadic value

fn as_ref_var(&self) -> Self::AsRefVar<'_>

Convert a reference to this variadic into a variadic of references.

let as_ref: var_type!(&u32, &String, &bool) =
    var_expr!(1_u32, "Hello".to_owned(), false).as_ref_var();

fn as_mut_var(&mut self) -> Self::AsMutVar<'_>

Convert an exclusive (mut) reference to this variadic into a variadic of exclusive (mut) references.

let as_mut: var_type!(&mut u32, &mut String, &mut bool) =
    var_expr!(1_u32, "Hello".to_owned(), false).as_mut_var();

fn iter_any_ref(&self) -> Self::IterAnyRef<'_>
where Self: 'static,

Iterate this variadic as &dyn Any references.

fn iter_any_mut(&mut self) -> Self::IterAnyMut<'_>
where Self: 'static,

Iterate this variadic as &mut dyn Any exclusive references.

fn into_option(self) -> Self::IntoOption

wrap all elements of the variadic in `Option``

fn into_singleton_vec(self) -> Self::IntoVec

wrap all elements of the variadic in a Vec

Provided Methods

fn len(&self) -> usize

The length of this variadic type

fn is_empty(&self) -> bool

Checks if this variadic type is empty.

Object Safety

This trait is not object safe.